helperメソッドの移管 #66
